api/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
websockets>=12.0
numpy>=1.24.0
sentence-transformers>=2.2.0
PyPDF2>=3.0.0
python-docx>=0.8.11
